[client] refactor/package.json
