+++ package.json
@@ -8,16 +8,30 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "axios": "^0.21.1",
     "autoprefixer": "^9.8.6",
     "browser-sync": "^2.27.3",
+    "dotenv": "^8.2.0",
+    "gulp": "^4.0.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-sass": "^5.0.0",
+    "gulp-sourcemaps": "^3.0.0",
     "gulp-postcss": "^9.0.0",
-    "react-intersection-observer": "^8.32.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.4",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^4.0.3",
+    "redux": "^4.1.0",
+    "redux-thunk": "^2.3.0",
+    "react-reveal": "^1.2.2",
     "react-select": "^4.3.1",
     "react-slick": "^0.28.1",
     "slick-carousel": "^1.8.1",
+    "react-intersection-observer": "^8.32.0",
+    "sass": "^1.35.1",
+    "scss": "^0.2.4",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -49,29 +63,12 @@
   },
   "devDependencies": {
     "@babel/highlight": "^7.14.5",
-    "axios": "^0.21.1",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
     "babel-plugin-transform-remove-console": "^6.9.4",
-    "dotenv": "^8.2.0",
     "eslint": "^7.29.0",
-    "gulp": "^4.0.2",
-    "gulp-concat": "^2.6.1",
-    "gulp-sass": "^5.0.0",
-    "gulp-sourcemaps": "^3.0.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-reveal": "^1.2.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "redux": "^4.1.0",
     "redux-devtools-extension": "^2.13.9",
-    "redux-logger": "^3.0.6",
-    "redux-saga": "^1.1.3",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.35.1",
-    "scss": "^0.2.4",
-    "tinycolor2": "^1.4.2"
+    "redux-logger": "^3.0.6"
   }
 }
