+++ src/page/Landing.js
@@ -57,7 +57,7 @@ function Landing() {
           <div className="landing__inner--second">
             <div className="desc-wrap desc-wrap--1">
               <Zoom duration={2300}>
-                <img className="desc1-img" src="desc01.png" alt="desc1" />
+                <img className="desc1-img" src="desc01.png" alt="desc1-img" />
 
                 <p className="desc1">
                   자주 들어가는 웹사이트
@@ -73,12 +73,12 @@ function Landing() {
                   <br />
                   여기저기 흩어져있는 북마크들...
                 </p>
-                <img className="desc2-img" src="desc002.png" alt="desc2" />
+                <img className="desc2-img" src="desc02.png" alt="desc2-img" />
               </Zoom>
             </div>
             <div className="desc-wrap desc-wrap--3">
               <Zoom>
-              <img className="desc3-img" src="desc03.png" alt="desc3" />
+              <img className="desc3-img" src="desc03.png" alt="desc3-img" />
                 <p className="desc3">
                   저장만 해두고
                   <br />
@@ -108,7 +108,7 @@ function Landing() {
             <div className="desc">
               <div className="desc__inner">
                 <Zoom duration={800} >
-                  <img src="collectvertical.gif" alt="collect-gif" />
+                  <img src="collectvertical.gif" alt="collectvertical-gif" />
                 </Zoom>
                 <div className="desc-text">
                   <Fade duration={2000}>
@@ -137,7 +137,7 @@ function Landing() {
               </Fade>
 
               <Pulse>
-                <img src="draganddrop.gif" alt="drag-drop-gif" />
+                <img src="draganddrop.gif" alt="draganddrop-gif" />
               </Pulse>
 
               <div className="desc-dragdrop__desc">
@@ -182,7 +182,7 @@ function Landing() {
               </div>
               <Fade duration={2000}>
                 <div className="alarm-img">
-                  <img src="./mockup.png" alt="alarm-img" />
+                  <img src="./alarm.png" alt="alarm-img" />
                 </div>
               </Fade>
             </div>
